requirements.txt
flask==3.0.3
gunicorn==23.0.0
colorgram.py
requests
pillow